BENTO-2319 Move SearchView to local-find package
