+++ packages/local-find/README.md
@@ -5,12 +5,27 @@ The Local Find package provides components relating to the Dashboard/Explore cas
   1. SearchBox – This is the case search autocomplete bar found on the Dashboard/Explore page. It provides a typeahead (autocomplete) wrapper with minimal configuration and lists the selected case filters.
   2. UploadModal – This is the modal that appears when a user clicks the "Upload Case Set" (by default) button. It handles TXT/CSV file uploads by default and passes the search entries to a search function.
 
+Additionally, this package provides a basic `SearchView` component that can be implemented on the Facet Search sidebar. See below for usage instructions.
+
 # Getting Started
 
 To import the components from LocalFind, use the following:
 
 ```javascript
-import { SearchBoxGenerator, UploadModalGenerator } from '@bento-core/local-find';
+import { SearchBoxGenerator, UploadModalGenerator, SearchView } from '@bento-core/local-find';
+
+/*
+ * Generate the SearchBox and UploadModal components here
+ */
+
+const FacetSearchView = (
+  <SearchView
+    classes={classes}
+    SearchBox={SearchBox}
+    UploadModal={UploadModal}
+    hidden={false}
+  />
+);
 ```
 
 Each of the components has their own respective README.md files. Please see the following links for detailed instructions on how to use them:
