+++ packages/local-find/src/index.js
@@ -4,6 +4,7 @@ export { SearchBoxGenerator } from './SearchBox/SearchBoxGenerator';
 export { default as SearchList } from './SearchBox/components/SearchList';
 export { DEFAULT_CONFIG_SEARCHBAR } from './SearchBox/config';
 export { default as LocalFindReducerGenerator } from './store/reducers/LocalFindReducerGenerator';
+export { default as SearchView } from './SearchView';
 export {
   updateAutocompleteData,
   updateUploadData,
